% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SY351SILAC.R
\docType{data}
\name{anno_dn_df}
\alias{anno_dn_df}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{.y.}{}
\item{group1}{}
\item{group2}{}
\item{p}{}
\item{p.adj}{}
\item{method}{}
\item{x.pos}{}
\item{log2r.shift.p.adj}{}
\item{y.position}{}
\item{motif}{}
\item{score}{}
\item{fg_match}{}
\item{adjusted_p.value}{}
\item{motifx_p.adj}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{SY351SILAC}
\link{anno_up_df}
\link{design.isotopeffect}
\link{dndf}
\link{phosphosites}
\link{proteinstats}
\link{proteinvolcano}
\link{psitestats}
\link{psitestats.grp}
\link{psitestats.protein.totals}
\link{targets}
\link{updf}
}
