% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SY351SILAC.R
\docType{data}
\name{design.isotopeffect}
\alias{design.isotopeffect}
\title{Design matrix for empirical Bayes moderated t-test analysis of +/- SY351 SILAC phosphopeptide ratios}
\format{
a \code{matrix} containing the following fields:
\describe{
}
}
\source{
The data comes from Old & Taatjes labs.
}
\description{
Design matrix for empirical Bayes moderated t-test analysis of +/- SY351 SILAC phosphopeptide ratios
}
\seealso{
\link{SY351SILAC}
\link{anno_dn_df}
\link{anno_up_df}
\link{dndf}
\link{phosphosites}
\link{proteinstats}
\link{proteinvolcano}
\link{psitestats}
\link{psitestats.grp}
\link{psitestats.protein.totals}
\link{targets}
\link{updf}
}
