% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SY351SILAC.R
\docType{data}
\name{phosphosites}
\alias{phosphosites}
\title{Phosphorylation site-level SILAC ratios.}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{null}{H:L SILAC ratio for DMSO:DMOS null condition}
\item{rep1}{H:L SILAC ratios for SY-351(Heavy):DMSO(Light) biological replicate comparison 1}
\item{rep2}{H:L SILAC ratios for SY-351(Heavy):DMSO(Light) biological replicate comparison 2}
\item{rep3LF}{Label swap H:L SILAC ratios for DMSO(Heavy):SY-351(Light) biological replicate comparison 3}
\item{Amino.acid}{}
\item{Charge}{}
\item{Reverse}{}
\item{Potential.contaminant}{}
\item{Multiplicity}{}
\item{Known.site}{}
\item{Origin}{}
\item{Regulatory.site}{}
\item{Regulatory.site.function}{}
\item{Regulatory.site.process}{}
\item{Regulatory.site.protInteract}{}
\item{Regulatory.site.otherInteract}{}
\item{Regulatory.site.notes}{}
\item{Motifs}{}
\item{Pfam.domains}{}
\item{active.site}{}
\item{binding.site}{}
\item{calcium.binding.region}{}
\item{chain}{}
\item{coiled.coil.region}{}
\item{compositionally.biased.region}{}
\item{cross.link}{}
\item{disulfide.bond}{}
\item{dna.binding.region}{}
\item{domain}{}
\item{glycosylation.site}{}
\item{helix}{}
\item{initiator.methionine}{}
\item{intramembrane.region}{}
\item{lipid.moiety.binding.region}{}
\item{metal.ion.binding.site}{}
\item{modified.residue}{}
\item{mutagenesis.site}{}
\item{non.consecutive.residues}{}
\item{non.standard.amino.acid}{}
\item{non.terminal.residue}{}
\item{nucleotide.phosphate.binding.region}{}
\item{peptide}{}
\item{propeptide}{}
\item{region.of.interest}{}
\item{repeat.}{}
\item{sequence.conflict}{}
\item{sequence.variant}{}
\item{short.sequence.motif}{}
\item{signal.peptide}{}
\item{site}{}
\item{splice.variant}{}
\item{strand}{}
\item{topological.domain}{}
\item{transit.peptide}{}
\item{transmembrane.region}{}
\item{turn}{}
\item{unsure.residue}{}
\item{zinc.finger.region}{}
\item{Localization.prob}{}
\item{PEP}{}
\item{Score}{}
\item{Delta.score}{}
\item{Score.for.localization}{}
\item{Mass.error..ppm.}{}
\item{Intensity}{}
\item{Intensity.L}{}
\item{Intensity.H}{}
\item{Position}{}
\item{Score.diff}{}
\item{Localization.prob.null}{}
\item{Score.diff.null}{}
\item{PEP.null}{}
\item{Score.null}{}
\item{Localization.prob.rep1}{}
\item{Score.diff.rep1}{}
\item{PEP.rep1}{}
\item{Score.rep1}{}
\item{Localization.prob.rep2}{}
\item{Score.diff.rep2}{}
\item{PEP.rep2}{}
\item{Score.rep2}{}
\item{Localization.prob.rep3LF}{}
\item{Score.diff.rep3LF}{}
\item{PEP.rep3LF}{}
\item{Score.rep3LF}{}
\item{Position.in.peptide}{}
\item{Ratio.H.L}{}
\item{Ratio.H.L___1}{}
\item{Ratio.H.L___2}{}
\item{Ratio.H.L___3}{}
\item{Ratio.H.L.normalized}{}
\item{Ratio.H.L.normalized___1}{}
\item{Ratio.H.L.normalized___2}{}
\item{Ratio.H.L.normalized___3}{}
\item{Ratio.H.L.unmod..pep.}{}
\item{Ratio.H.L.localized}{}
\item{Ratio.H.L.nmods}{}
\item{Ratio.H.L.variability....}{}
\item{Ratio.H.L.count}{}
\item{Ratio.H.L.iso.count}{}
\item{Ratio.H.L.type}{}
\item{Occupancy.L}{}
\item{Occupancy.H}{}
\item{Ratio.H.L.null}{}
\item{Ratio.H.L.null___1}{}
\item{Ratio.H.L.null___2}{}
\item{Ratio.H.L.null___3}{}
\item{Ratio.H.L.unmod..pep..null}{}
\item{Ratio.H.L.localized.null}{}
\item{Ratio.H.L.nmods.null}{}
\item{Ratio.H.L.variability.....null}{}
\item{Ratio.H.L.count.null}{}
\item{Ratio.H.L.iso.count.null}{}
\item{Ratio.H.L.type.null}{}
\item{Occupancy.L.null}{}
\item{Occupancy.H.null}{}
\item{Ratio.H.L.rep1}{}
\item{Ratio.H.L.rep1___1}{}
\item{Ratio.H.L.rep1___2}{}
\item{Ratio.H.L.rep1___3}{}
\item{Ratio.H.L.unmod..pep..rep1}{}
\item{Ratio.H.L.localized.rep1}{}
\item{Ratio.H.L.nmods.rep1}{}
\item{Ratio.H.L.variability.....rep1}{}
\item{Ratio.H.L.count.rep1}{}
\item{Ratio.H.L.iso.count.rep1}{}
\item{Ratio.H.L.type.rep1}{}
\item{Occupancy.L.rep1}{}
\item{Occupancy.H.rep1}{}
\item{Ratio.H.L.rep2}{}
\item{Ratio.H.L.rep2___1}{}
\item{Ratio.H.L.rep2___2}{}
\item{Ratio.H.L.rep2___3}{}
\item{Ratio.H.L.unmod..pep..rep2}{}
\item{Ratio.H.L.localized.rep2}{}
\item{Ratio.H.L.nmods.rep2}{}
\item{Ratio.H.L.variability.....rep2}{}
\item{Ratio.H.L.count.rep2}{}
\item{Ratio.H.L.iso.count.rep2}{}
\item{Ratio.H.L.type.rep2}{}
\item{Occupancy.L.rep2}{}
\item{Occupancy.H.rep2}{}
\item{Ratio.H.L.rep3LF}{}
\item{Ratio.H.L.rep3LF___1}{}
\item{Ratio.H.L.rep3LF___2}{}
\item{Ratio.H.L.rep3LF___3}{}
\item{Ratio.H.L.unmod..pep..rep3LF}{}
\item{Ratio.H.L.localized.rep3LF}{}
\item{Ratio.H.L.nmods.rep3LF}{}
\item{Ratio.H.L.variability.....rep3LF}{}
\item{Ratio.H.L.count.rep3LF}{}
\item{Ratio.H.L.iso.count.rep3LF}{}
\item{Ratio.H.L.type.rep3LF}{}
\item{Occupancy.L.rep3LF}{}
\item{Occupancy.H.rep3LF}{}
\item{Ratio.mod.base.L}{}
\item{Ratio.mod.base.H}{}
\item{Intensity.null}{}
\item{Intensity.L.null}{}
\item{Intensity.H.null}{}
\item{Ratio.mod.base.L.null}{}
\item{Ratio.mod.base.H.null}{}
\item{Intensity.rep1}{}
\item{Intensity.L.rep1}{}
\item{Intensity.H.rep1}{}
\item{Ratio.mod.base.L.rep1}{}
\item{Ratio.mod.base.H.rep1}{}
\item{Intensity.rep2}{}
\item{Intensity.L.rep2}{}
\item{Intensity.H.rep2}{}
\item{Ratio.mod.base.L.rep2}{}
\item{Ratio.mod.base.H.rep2}{}
\item{Intensity.rep3LF}{}
\item{Intensity.L.rep3LF}{}
\item{Intensity.H.rep3LF}{}
\item{Ratio.mod.base.L.rep3LF}{}
\item{Ratio.mod.base.H.rep3LF}{}
\item{Occupancy.null}{}
\item{Occupancy.rationull}{}
\item{Occupancy.error.scale.null}{}
\item{Occupancy.rep1}{}
\item{Occupancy.ratiorep1}{}
\item{Occupancy.error.scale.rep1}{}
\item{Occupancy.rep2}{}
\item{Occupancy.ratiorep2}{}
\item{Occupancy.error.scale.rep2}{}
\item{Occupancy.rep3LF}{}
\item{Occupancy.ratiorep3LF}{}
\item{Occupancy.error.scale.rep3LF}{}
\item{Proteins}{}
\item{Positions.within.proteins}{}
\item{Leading.proteins}{}
\item{Protein}{}
\item{Protein.names}{}
\item{Gene.names}{}
\item{Sequence.window}{}
\item{Fasta.headers}{}
\item{Diagnostic.peak}{}
\item{Modification.window}{}
\item{Peptide.window.coverage}{}
\item{Phospho..STY..Probabilities}{}
\item{Phospho..STY..Score.diffs}{}
\item{id}{}
\item{Positions}{}
\item{Peptide.IDs}{}
\item{Mod..peptide.IDs}{}
\item{Evidence.IDs}{}
\item{MS.MS.IDs}{}
\item{Best.localization.evidence.ID}{}
\item{Best.localization.MS.MS.ID}{}
\item{Best.localization.raw.file}{}
\item{Best.localization.scan.number}{}
\item{Best.score.evidence.ID}{}
\item{Best.score.MS.MS.ID}{}
\item{Best.score.raw.file}{}
\item{Best.score.scan.number}{}
\item{Best.PEP.evidence.ID}{}
\item{Best.PEP.MS.MS.ID}{}
\item{Best.PEP.raw.file}{}
\item{Best.PEP.scan.number}{}
\item{Unique.identifier}{}
\item{PhosphoSitePlus.window}{}
\item{PhosphoSitePlus.kinase}{}
\item{PhosphoSitePlus.kinase.uniprot}{}
\item{psite}{}
\item{numvalid}{}
\item{numvalid_trt}{}
\item{Gene.name}{}
\item{UniProt.names}{}
\item{Protein.name}{}
\item{Interpro.name}{}
\item{Prosite.name}{}
\item{proteinclass}{}
\item{Pfam}{}
\item{uniprot.acc.leading}{}
\item{gene.psite}{}
}
}
\source{
The data comes from________________________.
}
\description{
A data frame table (tibble) of the phosphorylation site-level SILAC ratios.
The data frame is a tbl_df or tibble - see dplyr package and tidyverse in general for info)
}
\seealso{
\link{SY351SILAC}
\link{anno_dn_df}
\link{anno_up_df}
\link{design.isotopeffect}
\link{dndf}
\link{proteinstats}
\link{proteinvolcano}
\link{psitestats}
\link{psitestats.grp}
\link{psitestats.protein.totals}
\link{targets}
\link{updf}
}
