% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SY351SILAC.R
\docType{data}
\name{proteinstats}
\alias{proteinstats}
\title{Table with statistics from differential protein expression analysis of protein SILAC ratios}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{id}{}
\item{null}{H:L SILAC ratio for DMSO:DMOS null condition}
\item{rep1}{H:L SILAC ratios for SY-351(Heavy):DMSO(Light) biological replicate comparison 1}
\item{rep2}{H:L SILAC ratios for SY-351(Heavy):DMSO(Light) biological replicate comparison 2}
\item{rep3LF}{Label swap H:L SILAC ratios for DMSO(Heavy):SY-351(Light) biological replicate comparison 3}
\item{Gene.name}{}
\item{uniprot.id}{}
\item{uniprot.acc}{}
\item{Protein.name}{}
\item{Corum}{}
\item{logFC}{}
\item{CI.L}{}
\item{CI.R}{}
\item{AveExpr}{}
\item{t}{}
\item{P.Value}{}
\item{adj.P.Val}{adjusted p-values using BH method}
\item{B}{log-ods for differential expression}
\item{s2.post}{}
\item{sigma}{}
\item{fdr.range}{}
\item{logp}{}
}
}
\source{
The data comes from________________________.
}
\description{
Table with statistics from differential protein expression analysis of protein SILAC ratios
}
\seealso{
\link{SY351SILAC}
\link{anno_dn_df}
\link{anno_up_df}
\link{design.isotopeffect}
\link{dndf}
\link{phosphosites}
\link{proteinvolcano}
\link{psitestats}
\link{psitestats.grp}
\link{psitestats.protein.totals}
\link{targets}
\link{updf}
}
