% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SY351SILAC.R
\docType{data}
\name{proteinvolcano}
\alias{proteinvolcano}
\title{Volcano plot for differential proteome analysis (trellis plot)}
\format{
a \code{trellis} containing the following fields:
\describe{
\item{formula}{}
\item{as.table}{}
\item{aspect.fill}{}
\item{legend}{}
\item{panel}{}
\item{page}{}
\item{layout}{}
\item{skip}{}
\item{strip}{}
\item{strip.left}{}
\item{xscale.components}{}
\item{yscale.components}{}
\item{axis}{}
\item{xlab}{}
\item{ylab}{}
\item{xlab.default}{}
\item{ylab.default}{}
\item{xlab.top}{}
\item{ylab.right}{}
\item{main}{}
\item{sub}{}
\item{x.between}{}
\item{y.between}{}
\item{par.settings}{}
\item{plot.args}{}
\item{lattice.options}{}
\item{par.strip.text}{}
\item{index.cond}{}
\item{perm.cond}{}
\item{condlevels}{}
\item{call}{}
\item{x.scales}{}
\item{y.scales}{}
\item{panel.args.common}{}
\item{panel.args}{}
\item{packet.sizes}{}
\item{x.limits}{}
\item{y.limits}{}
\item{x.used.at}{}
\item{y.used.at}{}
\item{x.num.limit}{}
\item{y.num.limit}{}
\item{aspect.ratio}{}
\item{prepanel.default}{}
\item{prepanel}{}
}
}
\source{
The data comes from________________________.
}
\description{
Volcano plot for differential proteome analysis (trellis plot)
}
\seealso{
\link{SY351SILAC}
\link{anno_dn_df}
\link{anno_up_df}
\link{design.isotopeffect}
\link{dndf}
\link{phosphosites}
\link{proteinstats}
\link{psitestats}
\link{psitestats.grp}
\link{psitestats.protein.totals}
\link{targets}
\link{updf}
}
