% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SY351SILAC.R
\docType{data}
\name{psitestats.grp}
\alias{psitestats.grp}
\title{Summary table for phosphorylation sites showing significant change.}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{Gene}{}
\item{UniprotAcc}{}
\item{PSite}{}
\item{Log2FC}{}
\item{AdjPval}{}
\item{PSeqWin}{}
\item{PhosphoSitePlus.kinase}{}
\item{Regulatory.site.function}{}
\item{proteinclass}{}
\item{numsig.down}{}
\item{numsig.up}{}
}
}
\source{
The data comes from________________________.
}
\description{
Summary table for individual phosphorylation sites showing significant change.
This table is generated by filtering the \link{psitestats} table using the following thresholds:
abs_null_log2ratio_thresh < 0.5
numvalid_trt_thresh > 2
fdr.thresh < 0.05
}
\seealso{
\link{SY351SILAC}
\link{anno_dn_df}
\link{anno_up_df}
\link{design.isotopeffect}
\link{dndf}
\link{phosphosites}
\link{proteinstats}
\link{proteinvolcano}
\link{psitestats}
\link{psitestats.protein.totals}
\link{targets}
\link{updf}
}
