% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SY351SILAC.R
\docType{data}
\name{psitestats.protein.totals}
\alias{psitestats.protein.totals}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{Uniprot_Acc}{}
\item{Gene}{}
\item{protein_function}{}
\item{high_conf_sites}{}
\item{numsig.down}{}
\item{min.log2fc}{}
\item{min.adj.pval}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{SY351SILAC}
\link{anno_dn_df}
\link{anno_up_df}
\link{design.isotopeffect}
\link{dndf}
\link{phosphosites}
\link{proteinstats}
\link{proteinvolcano}
\link{psitestats}
\link{psitestats.grp}
\link{targets}
\link{updf}
}
