% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SY351SILAC.R
\docType{package}
\name{SY351SILAC}
\alias{SY351SILAC}
\alias{SY351SILAC-package}
\title{Identification of CDK7 substrates with Syros' inhibitor SY-351 in HL-60 cells, using SILAC Phosphoproteomics}
\description{
R data package containing data for phosphorylation site and protein-level SILAC ratios (L:H)
from the MaxQuant PhosphoSTY.txt (and ProteinGroups.txt) files, which have also been processed with Perseus
to add site-level and protein-level annotations.
}
\details{
SY351SILAC
A data package for the phosphoproteomic and proteomic data sets published in Poss, ... Old, Taatjes and
scientists at Syros, using the Syros' CDK7 inhibitor SY-351. HL-60 cells were treated with inhibitor, followed by
SILAC (Light:Heavy) phosphoproteomics and proteomics analysis with TiO2 phosphopeptide enrichment and multidimensional
LC/MS/MS Fusion Orbitrap (Old Lab).

Use \code{data(package='SY351SILAC')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{anno_dn_df}
\link{anno_up_df}
\link{design.isotopeffect}
\link{dndf}
\link{phosphosites}
\link{proteinstats}
\link{proteinvolcano}
\link{psitestats}
\link{psitestats.grp}
\link{psitestats.protein.totals}
\link{targets}
\link{updf}
}
