% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SY351SILAC.R
\docType{data}
\name{targets}
\alias{targets}
\title{Descriptive data title}
\format{
a \code{data.frame} containing the following fields:
\describe{
\item{samples}{}
\item{Light}{}
\item{Heavy}{}
\item{samplenames}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{SY351SILAC}
\link{anno_dn_df}
\link{anno_up_df}
\link{design.isotopeffect}
\link{dndf}
\link{phosphosites}
\link{proteinstats}
\link{proteinvolcano}
\link{psitestats}
\link{psitestats.grp}
\link{psitestats.protein.totals}
\link{updf}
}
