% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SY351SILAC.R
\docType{data}
\name{updf}
\alias{updf}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{motifid}{}
\item{rowid}{}
\item{motif}{}
\item{seqwin}{}
\item{log2FC}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{SY351SILAC}
\link{anno_dn_df}
\link{anno_up_df}
\link{design.isotopeffect}
\link{dndf}
\link{phosphosites}
\link{proteinstats}
\link{proteinvolcano}
\link{psitestats}
\link{psitestats.grp}
\link{psitestats.protein.totals}
\link{targets}
}
